+++ src/errors/application-registration-not-exist.ts
