+++ src/utils/turtle-parser.ts
@@ -1,4 +1,4 @@
-import N3, { Prefixes, NamedNode } from "n3";
+import N3, { Prefixes } from "n3";
 import { Quad, DatasetCore } from "@rdfjs/types";
 const { Store, Parser, DataFactory } = N3;
